Move link wrapping into main script due to future click-to-play.
In order to support the Brave New World of "user interaction"-driven
autoplay rather than...well, autoplay, we don't want to have clicks
register on the link and cause it to change page. So in preparation
for this we'll move the wrap link into a separate function and will
eventually caller it later in the page load process.

Note: I've not tested with this exact path. :o