Rework to handle (primarily Chrome's) autoplay policy.
We attempt to detect if the audio context is suspended due to the
browser waiting for the person viewing the page to interact with it
before autoplaying audio. If this state is detected we display a
message prompt to interact with the page.

This also requires waiting until the GIF is about to  play before
wrapping the image with a link.

The implementation is getting pretty hacky with this change and
would be less so if we refactored to be a bit more object-orientated.

Maybe one day. :)